refactor: 마이페이지 시멘틱 태그 적용
