+++ src/components/mypage/MyInfo.jsx
@@ -3,7 +3,7 @@ import styled from "styled-components";
 const MyInfo = ({ children }) => {
   return <MyInfoField>{children}</MyInfoField>;
 };
-const MyInfoField = styled.div`
+const MyInfoField = styled.section`
   font-size: 14px;
   font-weight: 300;
 `;
